+++ R/elx_download_xml.R
@@ -16,7 +16,7 @@
 #' @export
 #' @examples
 #' \donttest{
-#' temploc <- paste(tempdir(), "elxnotice.xml", sep = "\\")
+#' temploc <- file.path(tempdir(), "elxnotice.xml")
 #' elx_download_xml(url = "http://publications.europa.eu/resource/celex/32022D0154",
 #'  file = temploc, notice = "object")
 #' unlink(temploc)
@@ -33,17 +33,17 @@ elx_download_xml <- function(url, file = paste(basename(url), ".xml", sep = ""),
   
   language <- paste(language_1,", ",language_2,";q=0.8, ",language_3,";q=0.7", sep = "")
   
-  if (stringr::str_detect(url,"celex.*[\\(|\\)|\\/]")){
-    
-    clx <- stringr::str_extract(url, "(?<=celex\\/).*") %>% 
-      stringr::str_replace_all("\\(","%28") %>% 
-      stringr::str_replace_all("\\)","%29") %>% 
-      stringr::str_replace_all("\\/","%2F")
-    
+  if (grepl("celex.*[\\(|\\)|\\/]", url)){
+
+    clx <- sub(".*celex/", "", url)
+    clx <- gsub("\\(", "%28", clx)
+    clx <- gsub("\\)", "%29", clx)
+    clx <- gsub("/", "%2F", clx)
+
     url <- paste("http://publications.europa.eu/resource/celex/",
                  clx,
                  sep = "")
-    
+
   }
   
   accept_header <- paste('application/xml; notice=',
@@ -77,5 +77,7 @@ elx_download_xml <- function(url, file = paste(basename(url), ".xml", sep = ""),
   utils::download.file(url = head$url,
                        destfile = file,
                        mode = mode)
+
+  return(invisible(file))
 
 }