+++ R/elx_label_eurovoc.R
@@ -6,7 +6,7 @@
 #' @param alt_labels If `TRUE`, results include comma-separated alternative labels in addition to the preferred label
 #' @param language Language in which to return the labels, in ISO 639 2-char code
 #' @return
-#' A `tibble` containing EuroVoc unique concept identifiers and labels.
+#' A data frame containing EuroVoc unique concept identifiers and labels.
 #' @export
 #' @examples
 #' \donttest{
@@ -26,7 +26,7 @@ elx_label_eurovoc <- function(uri_eurovoc = "", alt_labels = FALSE, language = "
 
   uniq_chunks <- split(uniq, ceiling(seq_along(uniq)/150))
 
-  uri_eurovoc_chunks <- purrr::map_chr(uniq_chunks,paste0,collapse = " ")
+  uri_eurovoc_chunks <- vapply(uniq_chunks, paste0, character(1), collapse = " ")
 
   query <- paste("PREFIX skos: <http://www.w3.org/2004/02/skos/core#>
   SELECT DISTINCT (group_concat(distinct ?subject;separator=',') as ?eurovoc)
@@ -48,7 +48,7 @@ elx_label_eurovoc <- function(uri_eurovoc = "", alt_labels = FALSE, language = "
   } GROUP BY ?subject",
                  sep = "")
 
-  out <- purrr::map_df(query,elx_run_query)
+  out <- do.call(rbind, lapply(query, elx_run_query))
 
   return(out)