+++ R/elx_run_query.R
@@ -30,11 +30,10 @@ elx_run_query <- function(query = "", endpoint = "http://publications.europa.eu/
   
   # if var created, continue
   if (!is.null(sparql_response)){
-    
+
     # parse response
-    sparql_response_parsed <- sparql_response %>% 
-      elx_parse_xml()
-    
+    sparql_response_parsed <- elx_parse_xml(sparql_response)
+
   } else {return(invisible(NULL))}
 
   # return
@@ -43,8 +42,6 @@ elx_run_query <- function(query = "", endpoint = "http://publications.europa.eu/
 }
 
 #' Fail http call gracefully
-#'
-#' @importFrom rlang .data
 #'
 #' @noRd
 #'
@@ -98,15 +95,13 @@ graceful_http <- function(remote_file, headers = NULL, body = NULL,
 #' An internal function to parse RDF/XML output from SPARQL queries.
 #'
 #' @noRd
-#' @importFrom rlang .data
 #'
 
 elx_parse_xml <- function(sparql_response = "", strip_uri = TRUE){
 
   # process XML response
-  res_binding <- sparql_response %>% 
-    xml2::read_xml() %>% 
-    xml2::xml_find_all("//d1:binding")
+  xml_doc <- xml2::read_xml(sparql_response)
+  res_binding <- xml2::xml_find_all(xml_doc, "//d1:binding")
 
   res_text <- xml2::xml_text(res_binding) 
 
@@ -115,44 +110,31 @@ elx_parse_xml <- function(sparql_response = "", strip_uri = TRUE){
   # eurovoc labels
   if (identical(unique(res_cols), c("eurovoc","labels"))){
 
-    out <- data.frame(res_cols, res_text) %>%
-      dplyr::mutate(is_work = dplyr::if_else(res_cols=="eurovoc", T, NA)) %>%
-      dplyr::group_by(.data$is_work) %>%
-      dplyr::mutate(triplet = dplyr::row_number(),
-                    triplet = dplyr::if_else(.data$is_work==T, .data$triplet, NA_integer_)) %>%
-      dplyr::ungroup() %>%
-      tidyr::fill(.data$triplet) %>%
-      dplyr::select(-"is_work") %>%
-      tidyr::pivot_wider(names_from = res_cols, values_from = res_text) %>%
-      dplyr::select(-"triplet")
+    df <- data.frame(res_cols = res_cols, res_text = res_text, stringsAsFactors = FALSE)
+    df$triplet <- cumsum(df$res_cols == "eurovoc")
+    out <- stats::reshape(df, idvar = "triplet", timevar = "res_cols", direction = "wide")
+    names(out) <- sub("^res_text\\.", "", names(out))
+    out$triplet <- NULL
 
-  } 
-  
+  }
+
   # regular result
   else {
 
-    out <- data.frame(res_cols, res_text) %>%
-      dplyr::mutate(is_work = dplyr::if_else(res_cols=="work", T, NA)) %>%
-      dplyr::group_by(.data$is_work) %>%
-      dplyr::mutate(triplet = dplyr::row_number(),
-                    triplet = dplyr::if_else(.data$is_work==T, .data$triplet, NA_integer_)) %>%
-      dplyr::ungroup() %>%
-      tidyr::fill(.data$triplet) %>%
-      dplyr::select(-"is_work") %>%
-      tidyr::pivot_wider(names_from = res_cols, values_from = res_text) %>%
-      dplyr::select(-"triplet")
+    df <- data.frame(res_cols = res_cols, res_text = res_text, stringsAsFactors = FALSE)
+    df$triplet <- cumsum(df$res_cols == "work")
+    out <- stats::reshape(df, idvar = "triplet", timevar = "res_cols", direction = "wide")
+    names(out) <- sub("^res_text\\.", "", names(out))
+    out$triplet <- NULL
 
   }
   
   # strip URI
   if (strip_uri == TRUE){
-    
-    out <- out |> 
-      dplyr::mutate(
-        dplyr::across(dplyr::everything(), 
-                      ~gsub("^http://publications.europa.eu/resource/cellar/|^http://publications.europa.eu/resource/authority/resource-type/","",.))
-      )
-    
+
+    uri_pattern <- "^http://publications.europa.eu/resource/cellar/|^http://publications.europa.eu/resource/authority/resource-type/"
+    out[] <- lapply(out, function(x) gsub(uri_pattern, "", x))
+
   }
 
   # end